Safari-609.1.11.1/LayoutTests/webgl/2.0.0/conformance2/rendering/uniform-block-buffer-size-expected.txt

CONSOLE MESSAGE: line 97: TypeError: Value is not a sequence
This test runs the WebGL Test listed below in an iframe and reports PASS or FAIL.

Test: ../../resources/webgl_test_files/conformance2/rendering/uniform-block-buffer-size.html
[ 1: PASS ] WebGL context exists
[ 2: FAIL ] successfullyParsed should be true (of type boolean). Was undefined (of type undefined).
[ FAIL ] 1 failures reported
